chore: resolving PR comments Updated mschema functions moved read_csv logic to bird_utils with multiple encodings
