+++ server/utilities/constants/bird_utils/response_messages.py
@@ -13,6 +13,8 @@ ERROR_MISSING_DB_ID = "Missing 'db_id' in item at index {index}."
 ERROR_PATH_NOT_EXIST = "Provided path does not exist: {dataset_directory}."
 ERROR_PATH_NOT_DIRECTORY = "Provided path is not a directory: {dataset_directory}."
 ERROR_EMPTY_BIRD_ITEMS_LIST = "The list of bird items is empty."
+ERROR_FAILED_TO_READ_CSV = "All encoding attempts failed for {file_path}"
 
 WARNING_DATABASE_DESCRIPTION_FILE_NOT_FOUND = "Database description file not found: {file_path}"
+WARNING_ENCODING_FAILED = "Failed with encoding {encoding}: {e}. Trying next..."
 WARNING_TABLE_DESCRIPTION_FILE_NOT_FOUND = "Table description file not found: {file_path}"
